requirements.txt
Flask
python-docx
spacy
Werkzeug
gunicorn
python-dotenv